% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dire_bonjour.R
\name{dire_bonjour}
\alias{dire_bonjour}
\title{Dire bonjour}
\usage{
dire_bonjour(entree)
}
\arguments{
\item{entree}{a string}
}
\value{
a string
}
\description{
Dire bonjour
}
\examples{
#' dire_bonjour("Sabine")

}
